Refactor code
This commit introduce a Repository class and moves the ProductId class
into a separate file for more clarity.
